hack for editable text - working on user branch
