+++ package.json
@@ -35,9 +35,15 @@
     "@babel/preset-env": "^7.12.1",
     "@babel/preset-flow": "^7.12.1",
     "@babel/preset-react": "^7.12.1",
+    "@emotion/react": "^11.10.5",
+    "@emotion/styled": "^11.10.5",
     "@material-ui/core": "^4.11.2",
     "@material-ui/icons": "^4.11.2",
     "@material-ui/lab": "^4.0.0-alpha.57",
+    "@mui/icons-material": "5.11.0",
+    "@mui/lab": "5.0.0-alpha.119",
+    "@mui/material": "5.11.8",
+    "@mui/styles": "5.11.7",
     "@typescript-eslint/eslint-plugin": "^4.14.1",
     "@unfoldingword/eslint-config": "1.3.0",
     "babel-loader": "^8.1.0",
@@ -66,7 +72,8 @@
     "scripture-resources-rcl": "^5.2.0",
     "style-loader": "^2.0.0",
     "tc-ui-toolkit": "^5.3.3",
-    "translation-helps-rcl": "1.47.0",
+    "translation-helps-rcl": "/Users/blm/Development/RCL/translation-helps-rcl/translation-helps-rcl-v3.3.6-alpha.tgz",
+    "use-deep-compare-effect": "1.8.1",
     "usfm-js": "3.4.2",
     "webpack": "^4.44.2",
     "word-aligner": "^1.0.0"
