+++ src/components/ScripturePane/ScripturePane.tsx
@@ -1,149 +1,10 @@
 import * as React from 'react'
-import _ from 'lodash'
-import {toUSFM} from 'usfm-js'
+import useDeepCompareEffect from 'use-deep-compare-effect'
 import { VerseObjects } from 'scripture-resources-rcl'
 import { ScriptureReference, VerseObjectsType } from '../../types'
 import { getResourceMessage } from '../../utils'
+import { getUsfmForVerseContent } from '../UsfmFileConversionHelpers'
 import { Container, Content } from './styled'
-
-/**
- * dive down into milestone to extract words and text
- * @param {Object} verseObject - milestone to parse
- * @return {string} text content of milestone
- */
-const parseMilestone = verseObject => {
-  let text = verseObject.text || '';
-  let wordSpacing = '';
-  const length = verseObject.children ? verseObject.children.length : 0;
-
-  for (let i = 0; i < length; i++) {
-    let child = verseObject.children[i];
-
-    switch (child.type) {
-    case 'word':
-      text += wordSpacing + child.text;
-      wordSpacing = ' ';
-      break;
-
-    case 'milestone':
-      text += wordSpacing + parseMilestone(child);
-      wordSpacing = ' ';
-      break;
-
-    default:
-      if (child.text) {
-        text += child.text;
-        const lastChar = text.substr(-1);
-
-        if ((lastChar !== ',') && (lastChar !== '.') && (lastChar !== '?') && (lastChar !== ';')) { // legacy support, make sure padding before word
-          wordSpacing = '';
-        }
-      }
-      break;
-    }
-  }
-  return text;
-};
-
-/**
- * get text from word and milestone markers
- * @param {Object} verseObject - to parse
- * @param {String} wordSpacing - spacing to use before next word
- * @return {*} new verseObject and word spacing
- */
-const replaceWordsAndMilestones = (verseObject, wordSpacing) => {
-  let text = '';
-
-  if (verseObject.type === 'word') {
-    text = wordSpacing + verseObject.text;
-  } else if (verseObject.type === 'milestone') {
-    text = wordSpacing + parseMilestone(verseObject);
-  }
-
-  if (text) { // replace with text object
-    verseObject = {
-      type: 'text',
-      text,
-    };
-    wordSpacing = ' ';
-  } else {
-    wordSpacing = ' ';
-
-    if (verseObject.nextChar) {
-      wordSpacing = ''; // no need for spacing before next word if this item has it
-    } else if (verseObject.text) {
-      const lastChar = verseObject.text.substr(-1);
-
-      if (![',', '.', '?', ';'].includes(lastChar)) { // legacy support, make sure padding before next word if punctuation
-        wordSpacing = '';
-      }
-    }
-
-    if (verseObject.children) { // handle nested
-      const verseObject_ = _.cloneDeep(verseObject);
-      let wordSpacing_ = '';
-      const length = verseObject.children.length;
-
-      for (let i = 0; i < length; i++) {
-        const flattened =
-          replaceWordsAndMilestones(verseObject.children[i], wordSpacing_);
-        wordSpacing_ = flattened.wordSpacing;
-        verseObject_.children[i] = flattened.verseObject;
-      }
-      verseObject = verseObject_;
-    }
-  }
-  return { verseObject, wordSpacing };
-};
-
-/**
- * converts verse from verse objects to USFM string
- * @param verseData
- * @return {string}
- */
-export function convertVerseDataToUSFM(verseData) {
-  const outputData = {
-    'chapters': {},
-    'headers': [],
-    'verses': { '1': verseData },
-  };
-  const USFM = toUSFM(outputData, { chunk: true });
-  const split = USFM.split('\\v 1');
-
-  if (split.length > 1) {
-    let content = split[1];
-
-    if (content.substr(0, 1) === ' ') { // remove space separator
-      content = content.substr(1);
-    }
-    return content;
-  }
-  return ''; // error on JSON to USFM
-}
-
-/**
- * @description convert verse from verse objects to USFM string, removing milestones and word markers
- * @param {Object|Array} verseData
- * @return {String}
- */
-const getUsfmForVerseContent = (verseData) => {
-  if (verseData.verseObjects) {
-    let wordSpacing = '';
-    const flattenedData = [];
-    const length = verseData.verseObjects.length;
-
-    for (let i = 0; i < length; i++) {
-      const verseObject = verseData.verseObjects[i];
-      const flattened = replaceWordsAndMilestones(verseObject, wordSpacing);
-      wordSpacing = flattened.wordSpacing;
-      flattenedData.push(flattened.verseObject);
-    }
-    verseData = { // use flattened data
-      verseObjects: flattenedData,
-    };
-  }
-  return convertVerseDataToUSFM(verseData);
-};
 
 interface Props {
   /** current reference **/
@@ -174,8 +35,10 @@ interface Props {
   translate: Function;
   /** true if in edit mode */
   editing: boolean;
-  /** to set edit mode */
+  /** callback to set edit mode */
   setEditing: Function;
+  /** callback to set that verse has changed */
+  setVerseChanged: Function;
 }
 
 const MessageStyle = {
@@ -208,7 +71,9 @@ function ScripturePane({
   translate,
   editing,
   setEditing,
+  setVerseChanged,
 } : Props) {
+  const [initialVerseText, setInitialVerseText] = React.useState(null)
   const resourceMsg = getResourceMessage(resourceStatus, server, resourceLink, isNT)
   const { chapter, verse } = reference
   direction = direction || 'ltr'
@@ -221,6 +86,19 @@ function ScripturePane({
   contentStyle = contentStyle || {
     fontFamily: 'Noto Sans',
     fontSize: '100%',
+  }
+
+  useDeepCompareEffect(() => {
+    const verseText = getUsfmForVerseContent({ verseObjects })
+    setInitialVerseText(verseText)
+  }, [verseObjects])
+
+  function onTextChange(event: React.ChangeEvent<HTMLTextAreaElement>) {
+    // console.log(`onTextChange`, event)
+    const newText = event?.target?.value
+    const changed = newText !== initialVerseText
+    setVerseChanged(changed)
+    console.log(`onTextChange: new text ${changed ? 'changed' : 'unchanged'}: `, newText)
   }
 
   return (
@@ -238,7 +116,11 @@ function ScripturePane({
           }}
           >
             {editing ?
-              <textarea defaultValue={getUsfmForVerseContent({ verseObjects })} />
+              <textarea
+                defaultValue={initialVerseText}
+                onChange={onTextChange}
+                style={{ height: '60%', width: '300px' }}
+              />
               :
               <VerseObjects
                 verseObjects={verseObjects}
